Change the "infer_decoupled" JSON format
Essentially there was a bug that necessitated changing theh output
format: if you have a DecoupledIO that just wraps a Bits then you get
the output name of "BASE_bits".  This conflicts with my assumption that
stripped "BASE_bits_" from all the IOs, so I had to change the output
format.

The JSON output format now contains the full name of all the ports in
the "bits" array.