Add "NEXT_STRUCTURE" to the JSON dumper
The general idea here is that sometimes you want to name a structure,
but can't provide the contents of that structure until later (for
example, if you're calling into some other function).  To allow whatever
produces the structures later to be internally consistant, it must
start with a BEGIN_STRUCTURE.  Thus, you name the structure and provied
NEXT_STRUCTURE to the stream, which will then expect a BEGIN_STRUCTURE
as the next token (and disallow anything else).