Add a Verilog emitter and canonicalizer binary
I need to be able to emit Verilog from PCAD for other tools to use.
This adds another dumper (like the JSON dumper) that formats a HDL AST
as Verilog.  There's a simple test case that runs on all circuits:
canonicalizing a Verilog file must be idempotent.

There's also some hand-written test cases, but they need to be expanded
upon.