Add "someptr" to simple_match
I use shared pointers for _everything_, but simple_match seems to want
to convert them to reference all the time.  While this might be the
right thing to do (it's probably broken to have shared_ptr everywhere in
my interfaces), I don't care.

someptr is just like some, but it keeps the shared pointer when calling
the match function.  There's no support for someptr() because I can just
stick the auto on the other side.