Part of a memory compiler
This passes data all the way through to the HDL AST to print out
compiled memory macros but doesn't actually fill out any of the modules.
I kind of want to do a decent job at the RTL IR (since the HDL AST is
a mess) and I'm too tired to deal with it now.