Have the macro compiler split memories based on width
This is a very simple macro compiler right now: it'll only match
identical port sets, only handles single-port memories, and only handles
a single memory macro as the target.

That said, I'm pretty sure it produces correct Verilog for this use case
so I'm commiting it with a test case.