Add the "--syn-flops" argument to the macro compiler
The macro compiler defaults to emitting flop-based memories when there
isn't a cooresponding macro that can implement the desired memory.  We
use these in two ways: during RTL simulation to provide
technology-independent memories, and at synthesis to ensure the circuit
is always fully implemented.  The result is that I want two different
implementations to come out: one that's safe (does randomization and
such), and one that's synthesizable.  I know this is kind of awkward,
but we'll just have to trust the generator for now.

This is kind of funky: I don't like having boolean arguments to my
functions, and there's going to end up being a lot of them if I keep
going this way.