Always emit 1-bit wires as "[0:0]"
I don't actually know Verilog, so I'm not sure what the right answer is
here.  VCS throws a warning when I "[0]" a wire that's declared as
"wire", but doesn't when I declare it as "wire [0:0]".  As far as I
know, aside from part select these are the same statement, so I'm just
always going to emit "wire [0:0]".

This requires a lot of test changes.