Add an early out (for speed) in the memory compiler
This applies a simple cost function that acts as a lower bound of the
cost of a memory, which can then be used to eliminate unnecessary calls
to the compiler.  On technologies with lots of tiny SRAMs, the complier
spends tons of time trying to target memories that are obviously bad,
this fixes the problem.